fix: All bugs fixed. Compilation is successfull.

+++ backend/src/main/java/edu/iztech/utms/g02/utms_app/api/application/controller/ApplicationController.java
@@ -7,16 +7,15 @@ import edu.iztech.utms.g02.utms_app.api.application.dto.YdyoReviewRequest;
 
 import edu.iztech.utms.g02.utms_app.bl.application.ApplicationService;
 
-import jakarta.validation.Valid;
+
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import org.springframework.security.access.prepost.PreAuthorize;      //??????????
 import org.springframework.web.multipart.MultipartFile;                //??????????
 
-import java.util.Map;
+
 import java.util.List;
 
 @RestController
@@ -35,7 +34,7 @@ public class ApplicationController {
     @PostMapping
     public ResponseEntity<ApplicationResponse> createApplication(@RequestBody ApplicationCreateRequest req) {
         // Not: Gerçek projede userId'yi SecurityContext'ten (token'dan) alırsın
-        Long currentUserId = 1L; // Şimdilik temsili
+        Integer currentUserId = 1; // Şimdilik temsili // ******************************
         ApplicationResponse response = applicationService.create(currentUserId, req);
         return ResponseEntity.ok(response);
     }
@@ -47,7 +46,7 @@ public class ApplicationController {
 
     @PreAuthorize("hasRole('STUDENT')") 
     @PatchMapping("/{id}/submit")
-    public ResponseEntity<String> submitApplication(@PathVariable Long id) {
+    public ResponseEntity<String> submitApplication(@PathVariable Integer id) {
         // Öğrencinin sadece kendi başvurusunu gönderebilmesi kontrolü Service katmanında yapılır
         applicationService.submit(id, null);
         return ResponseEntity.ok("Başvuru başarıyla gönderildi.");
@@ -69,7 +68,7 @@ public class ApplicationController {
 
     @PreAuthorize("hasAnyRole('STUDENT', 'OIDB', 'YDYO', 'FACULTY', 'DEAN')") 
     @GetMapping("/{id}")
-    public ResponseEntity<ApplicationResponse> getApplicationById(@PathVariable Long id) {
+    public ResponseEntity<ApplicationResponse> getApplicationById(@PathVariable Integer id) {
         ApplicationResponse response = applicationService.getApplicationById(id);
         return ResponseEntity.ok(response);
     }
@@ -78,7 +77,7 @@ public class ApplicationController {
 
     @PreAuthorize("hasRole('OIDB')")
     @PatchMapping("/{id}/oidb-review")
-    public ResponseEntity<String> reviewByOidb(@PathVariable Long id, @RequestBody OidbReviewRequest req) {
+    public ResponseEntity<String> reviewByOidb(@PathVariable Integer id, @RequestBody OidbReviewRequest req) {
         applicationService.processOidbReview(id, req);
         return ResponseEntity.ok("OIDB incelemesi kaydedildi.");
     }
@@ -87,7 +86,7 @@ public class ApplicationController {
 
     @PreAuthorize("hasRole('YDYO')")
     @PatchMapping("/{id}/ydyo-review")
-    public ResponseEntity<String> reviewByYdyo(@PathVariable Long id, @RequestBody YdyoReviewRequest req) {
+    public ResponseEntity<String> reviewByYdyo(@PathVariable Integer id, @RequestBody YdyoReviewRequest req) {
         applicationService.processYdyoReview(id, req);
         return ResponseEntity.ok("YDYO incelemesi kaydedildi.");
     }
@@ -96,7 +95,7 @@ public class ApplicationController {
     @PreAuthorize("hasRole('STUDENT')")
     @PostMapping(value = "/{id}/documents", consumes = "multipart/form-data")       // consumes ??
     public ResponseEntity<String> uploadDocument(
-            @PathVariable Long id, 
+            @PathVariable Integer id, 
             @RequestParam("file") MultipartFile file) {
         
         applicationService.uploadDocument(id, file);
